Moved WaveSpawner to its own file
